user_registration_type edited

+++ database/seeds/SettingTableSeeder.php
@@ -99,7 +99,7 @@ class SettingTableSeeder extends Seeder
          * */
         Setting::create([
             'attribute_key'               => 'user_registration_type',
-            'attribute_value'             => Setting::PUBLIC,
+            'attribute_value'             => (int) Setting::PUBLIC,
         ]);
 
         Setting::create([
